HealthCoachBackend/requirements.txt
fastapi
uvicorn
requests
spacy
nltk
scikit-learn
psycopg2-binary
SQLAlchemy
python-dotenv
alembic